Add random color circle when no image
